+++ bin/deploy/src/local.rs
@@ -112,7 +112,8 @@ fn build_validators(
             metrics_port,
             max_propose_bytes: args.max_propose_bytes,
             max_pool_bytes: args.max_pool_bytes,
-            page_cache_bytes: args.page_cache_bytes,
+            state_page_cache_bytes: args.state_page_cache_bytes,
+            archive_page_cache_bytes: args.archive_page_cache_bytes,
             public_key_cache_size: args.public_key_cache_size,
             traces: 0.0,
             bootstrappers: bootstrappers.clone(),
@@ -187,7 +188,8 @@ fn build_secondaries(
             metrics_port,
             max_propose_bytes: args.max_propose_bytes,
             max_pool_bytes: args.max_pool_bytes,
-            page_cache_bytes: args.page_cache_bytes,
+            state_page_cache_bytes: args.state_page_cache_bytes,
+            archive_page_cache_bytes: args.archive_page_cache_bytes,
             public_key_cache_size: args.public_key_cache_size,
             traces: 0.0,
             bootstrappers: bootstrappers.clone(),
@@ -415,7 +417,8 @@ mod tests {
             public_key_cache_size: default_public_key_cache_size(),
             max_propose_bytes: default_max_propose_bytes(),
             max_pool_bytes: default_max_pool_bytes(),
-            page_cache_bytes: default_page_cache_bytes(),
+            state_page_cache_bytes: default_page_cache_bytes(),
+            archive_page_cache_bytes: default_page_cache_bytes(),
             startup: StartupModeConfig::MarshalSync,
             spammer,
             spammer_accounts: 10,
