+++ bin/deploy/src/main.rs
@@ -124,11 +124,14 @@ pub(crate) struct GenerateArgs {
     /// Maximum mempool size in bytes.
     #[arg(long = "max-pool-bytes", default_value_t = default_max_pool_bytes())]
     max_pool_bytes: usize,
-    /// Capacity in bytes of each validator's two storage page caches. Must
+    /// Capacity in bytes of each validator's state QMDB page cache. Must
     /// hold the state journal's working set: 512 MiB thrashed once the live
     /// account set passed ~2M and build/verify doubled on cache misses.
-    #[arg(long = "page-cache-bytes", default_value_t = default_page_cache_bytes())]
-    page_cache_bytes: usize,
+    #[arg(long = "state-page-cache-bytes", default_value_t = default_page_cache_bytes())]
+    state_page_cache_bytes: usize,
+    /// Capacity in bytes of each validator's archive/journal page cache.
+    #[arg(long = "archive-page-cache-bytes", default_value_t = default_page_cache_bytes())]
+    archive_page_cache_bytes: usize,
     /// Startup sync mode for the generated validators.
     #[arg(long, value_enum, default_value_t = StartupModeConfig::MarshalSync)]
     startup: StartupModeConfig,
@@ -394,8 +397,10 @@ pub(crate) struct ValidatorConfig {
     max_propose_bytes: usize,
     /// Maximum mempool size in bytes.
     max_pool_bytes: usize,
-    /// Capacity in bytes of each of the engine's two storage page caches.
-    page_cache_bytes: usize,
+    /// Capacity in bytes of the engine's state QMDB page cache.
+    state_page_cache_bytes: usize,
+    /// Capacity in bytes of the engine's archive/journal page cache.
+    archive_page_cache_bytes: usize,
     /// Capacity of the decompressed public key cache.
     #[serde(default = "default_public_key_cache_size")]
     public_key_cache_size: usize,
