+++ bin/deploy/src/remote.rs
@@ -166,7 +166,8 @@ fn build_validators(
             metrics_port: METRICS_PORT,
             max_propose_bytes: args.max_propose_bytes,
             max_pool_bytes: args.max_pool_bytes,
-            page_cache_bytes: args.page_cache_bytes,
+            state_page_cache_bytes: args.state_page_cache_bytes,
+            archive_page_cache_bytes: args.archive_page_cache_bytes,
             public_key_cache_size: args.public_key_cache_size,
             traces: remote.traces,
             bootstrappers: bootstrappers.clone(),
@@ -221,7 +222,8 @@ fn build_secondaries(
             metrics_port: METRICS_PORT,
             max_propose_bytes: args.max_propose_bytes,
             max_pool_bytes: args.max_pool_bytes,
-            page_cache_bytes: args.page_cache_bytes,
+            state_page_cache_bytes: args.state_page_cache_bytes,
+            archive_page_cache_bytes: args.archive_page_cache_bytes,
             public_key_cache_size: args.public_key_cache_size,
             traces: remote.traces,
             bootstrappers: bootstrappers.clone(),
@@ -510,7 +512,8 @@ mod tests {
             public_key_cache_size: default_public_key_cache_size(),
             max_propose_bytes: default_max_propose_bytes(),
             max_pool_bytes: default_max_pool_bytes(),
-            page_cache_bytes: default_page_cache_bytes(),
+            state_page_cache_bytes: default_page_cache_bytes(),
+            archive_page_cache_bytes: default_page_cache_bytes(),
             startup: StartupModeConfig::MarshalSync,
             spammer: false,
             spammer_accounts: 10,
@@ -581,7 +584,8 @@ mod tests {
                 metrics_port: 9090,
                 max_propose_bytes: default_max_propose_bytes(),
                 max_pool_bytes: default_max_pool_bytes(),
-                page_cache_bytes: default_page_cache_bytes(),
+                state_page_cache_bytes: default_page_cache_bytes(),
+                archive_page_cache_bytes: default_page_cache_bytes(),
                 public_key_cache_size: default_public_key_cache_size(),
                 traces: 0.0,
                 bootstrappers: Vec::new(),
