+++ crates/engine/src/tests/mod.rs
@@ -296,7 +296,8 @@ impl EngineDefinition for TestEngineDefinition {
                     namespace: ENGINE_NAMESPACE.to_vec(),
                     // Small: simulation state is tiny and large caches slow
                     // deterministic runs.
-                    page_cache_bytes: 32 * 1024 * 1024,
+                    state_page_cache_bytes: 32 * 1024 * 1024,
+                    archive_page_cache_bytes: 32 * 1024 * 1024,
                     output,
                     share,
                     input,
